+++ tests/test_hubspot_pagination.py
@@ -1,9 +1,12 @@
-import tap_tester.connections as connections
-import tap_tester.menagerie   as menagerie
-import tap_tester.runner      as runner
 from datetime import datetime
 from datetime import timedelta
 import time
+
+import tap_tester.connections as connections
+import tap_tester.menagerie   as menagerie
+import tap_tester.runner      as runner
+from tap_tester.logger import LOGGER
+
 from client import TestClient
 from base import HubspotBaseTest
 
@@ -55,24 +58,22 @@ class TestHubspotPagination(HubspotBaseTest):
 
             # check if we exceed the pagination limit
             under_target = limits[stream] + 1 - len(existing_records[stream])
-            print(f'under_target = {under_target} for {stream}')
+            LOGGER.info(f'under_target = {under_target} for {stream}')
 
             # if we do not exceed the limit generate more data so that we do
             if under_target > 0 :
-                print(f"need to make {under_target} records for {stream} stream")
+                LOGGER.info(f"need to make {under_target} records for {stream} stream")
                 if stream in {'subscription_changes', 'emails_events'}:
                     test_client.create(stream, subscriptions=existing_records[stream], times=under_target)
+                elif stream == 'contacts_by_company':
+                    test_client.create(stream, company_ids, times=under_target)
                 else:
                     for i in range(under_target):
                         # create records to exceed limit
-                        if stream == 'contacts_by_company':
-                            # company_ids = [record["company-id"] for record in existing_records[stream]]
-                            test_client.create(stream, company_ids)
-                        else:
-                            test_client.create(stream)
+                        test_client.create(stream)
 
         setup_end = time.perf_counter()
-        print(f"Test Client took about {str(setup_end-setup_start).split('.')[0]} seconds")
+        LOGGER.info(f"Test Client took about {str(setup_end-setup_start).split('.')[0]} seconds")
 
     def streams_to_test(self):
         """
@@ -84,6 +85,9 @@ class TestHubspotPagination(HubspotBaseTest):
             if limit
         }
         streams_to_test = streams_with_page_limits.difference({
+            # updates for contacts_by_company do not get processed quickly or consistently
+            # via Hubspot API, unable to guarantee page limit is exceeded
+            'contacts_by_company',
             'email_events',
             'subscription_changes', # BUG_TDL-14938 https://jira.talendforge.org/browse/TDL-14938
         })